clean up links work
