+++ src/views/editor/blocks/markdown.tsx
@@ -1,0 +1,171 @@
+import Prism from "prismjs";
+import React, { useState, useCallback, useMemo, useEffect, useRef } from "react";
+import { Slate, Editable, withReact, ReactEditor, RenderElementProps, useSlate } from "slate-react";
+import { Text, createEditor, Node, Element as SlateElement, Editor, Selection } from "slate";
+import { withHistory } from "slate-history";
+import { css } from "emotion";
+// import { isImageElement, isLinkElement, ImageElement, LinkElement } from './util';
+// import { Link } from './blocks/links';
+// import { insertLink, removeLink } from './blocks/links';
+import { TextInput, IconButton, Pane, TextInputField, Button } from 'evergreen-ui'
+
+Prism.languages.markdown = Prism.languages.extend("markup", {});
+(Prism.languages as any).insertBefore("markdown", "prolog", {
+  blockquote: { pattern: /^>(?:[\t ]*>)*/m, alias: "punctuation" },
+  code: [
+    { pattern: /^(?: {4}|\t).+/m, alias: "keyword" },
+    { pattern: /``.+?``|`[^`\n]+`/, alias: "keyword" },
+  ],
+  title: [
+    {
+      pattern: /\w+.*(?:\r?\n|\r)(?:==+|--+)/,
+      alias: "important",
+      inside: { punctuation: /==+$|--+$/ },
+    },
+    {
+      pattern: /(^\s*)#+.+/m,
+      lookbehind: !0,
+      alias: "important",
+      inside: { punctuation: /^#+|#+$/ },
+    },
+  ],
+  hr: {
+    pattern: /(^\s*)([*-])([\t ]*\2){2,}(?=\s*$)/m,
+    lookbehind: !0,
+    alias: "punctuation",
+  },
+  list: {
+    pattern: /(^\s*)(?:[*+-]|\d+\.)(?=[\t ].)/m,
+    lookbehind: !0,
+    alias: "punctuation",
+  },
+  "url-reference": {
+    pattern:
+      /!?\[[^\]]+\]:[\t ]+(?:\S+|<(?:\\.|[^>\\])+>)(?:[\t ]+(?:"(?:\\.|[^"\\])*"|'(?:\\.|[^'\\])*'|\((?:\\.|[^)\\])*\)))?/,
+    inside: {
+      variable: { pattern: /^(!?\[)[^\]]+/, lookbehind: !0 },
+      string: /(?:"(?:\\.|[^"\\])*"|'(?:\\.|[^'\\])*'|\((?:\\.|[^)\\])*\))$/,
+      punctuation: /^[\[\]!:]|[<>]/,
+    },
+    alias: "url",
+  },
+  bold: {
+    pattern: /(^|[^\\])(\*\*|__)(?:(?:\r?\n|\r)(?!\r?\n|\r)|.)+?\2/,
+    lookbehind: !0,
+    inside: { punctuation: /^\*\*|^__|\*\*$|__$/ },
+  },
+  italic: {
+    pattern: /(^|[^\\])([*_])(?:(?:\r?\n|\r)(?!\r?\n|\r)|.)+?\2/,
+    lookbehind: !0,
+    inside: { punctuation: /^[*_]|[*_]$/ },
+  },
+  url: {
+    pattern:
+      /!?\[[^\]]+\](?:\([^\s)]+(?:[\t ]+"(?:\\.|[^"\\])*")?\)| ?\[[^\]\n]*\])/,
+    inside: {
+      variable: { pattern: /(!?\[)[^\]]+(?=\]$)/, lookbehind: !0 },
+      string: { pattern: /"(?:\\.|[^"\\])*"(?=\)$)/ },
+    },
+  },
+});
+(Prism.languages.markdown.bold as any).inside.url = Prism.util.clone(
+  Prism.languages.markdown.url,
+);
+(Prism.languages.markdown.italic as any).inside.url = Prism.util.clone(
+  Prism.languages.markdown.url,
+);
+(Prism.languages.markdown.bold as any).inside.italic = Prism.util.clone(
+  Prism.languages.markdown.italic,
+);
+(Prism.languages.markdown.italic as any).inside.bold = Prism.util.clone(
+  Prism.languages.markdown.bold,
+);
+
+
+export const useDecorateMarkdown = () => {
+  return useCallback(([node, path]) => {
+    const ranges: any = [];
+
+    if (!Text.isText(node)) {
+      return ranges;
+    }
+
+    const getLength = (token: any) => {
+      if (typeof token === "string") {
+        return token.length;
+      } else if (typeof token.content === "string") {
+        return token.content.length;
+      } else {
+        return token.content.reduce((l: any, t: any) => l + getLength(t), 0);
+      }
+    };
+
+    const tokens = Prism.tokenize(node.text, Prism.languages.markdown);
+    let start = 0;
+
+    for (const token of tokens) {
+      const length = getLength(token);
+      const end = start + length;
+
+      if (typeof token !== "string") {
+        ranges.push({
+          [token.type]: true,
+          anchor: { path, offset: start },
+          focus: { path, offset: end },
+        });
+      }
+
+      start = end;
+    }
+
+    return ranges;
+  }, []);
+}
+
+export const MarkdownLeaf = ({ attributes, children, leaf }: any) => {
+  return (
+    <span
+      {...attributes}
+      className={css`
+        font-weight: ${leaf.bold && "bold"};
+        font-style: ${leaf.italic && "italic"};
+        text-decoration: ${leaf.underlined && "underline"};
+        ${leaf.title &&
+        css`
+            display: inline-block;
+            font-weight: bold;
+            font-size: 20px;
+            margin: 20px 0 10px 0;
+          `}
+        ${leaf.list &&
+        css`
+            padding-left: 10px;
+            font-size: 20px;
+            line-height: 10px;
+          `}
+        ${leaf.hr &&
+        css`
+            display: block;
+            text-align: center;
+            border-bottom: 2px solid #ddd;
+          `}
+        ${leaf.blockquote &&
+        css`
+            display: inline-block;
+            border-left: 2px solid #ddd;
+            padding-left: 10px;
+            color: #aaa;
+            font-style: italic;
+          `}
+        ${leaf.code &&
+        css`
+            font-family: monospace;
+            background-color: #eee;
+            padding: 3px;
+          `}
+      `}
+    >
+      {children}
+    </span>
+  );
+};
