+++ src/views/editor/withHelpers.tsx
@@ -3,21 +3,26 @@
 import { Text, Transforms, Node as SlateNode, Range, Path as SlatePath, createEditor, Descendant, Editor, Element as SlateElement } from 'slate'
 import { ReactEditor } from 'slate-react';
 
-// todo: centralize these utilities
+// todo: centralize these utilities -- they are also used in a few other places and can get out of sync
+// which would produce weird bugs! 
 import unified from "unified";
 import markdown from "remark-parse";
 import remarkGfm from 'remark-gfm'
 import { remarkToSlate, slateToRemark, mdastToSlate } from "remark-slate-transformer";
 const parser = unified().use(markdown).use(remarkGfm as any)
 import { isTypedElement, isLinkElement } from './util';
-import { insertLink } from './blocks/links';
+import { insertLink, urlMatcher } from './blocks/links';
 
-// todo: move to links
-const urlRegex = /[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_\+.~#?&//=]*)?/gi;
-const urlMatcher = new RegExp(urlRegex);
 
-export const withImages = (editor: ReactEditor) => {
-  const { insertData, isVoid, insertBreak, insertText, normalizeNode, isInline } = editor
+/**
+ * Image and link helpers. Maybe more.
+ * 
+ * Will look at refactoring once I finish the first phase of wysiwyg work and understand it better.
+ * @param editor 
+ * @returns 
+ */
+export const withHelpers = (editor: ReactEditor) => {
+  const { isVoid, normalizeNode, isInline } = editor
  
   // If the element is an image type, make it non-editable
   // https://docs.slatejs.org/concepts/02-nodes#voids
@@ -104,167 +109,11 @@ export const withImages = (editor: ReactEditor) => {
   return editor
 }
 
-// const insertImage = (editor, url) => {
-//   const text = { text: '' }
-//   const image: ImageElement = { type: 'image', url, children: [text] }
-//   Transforms.insertNodes(editor, image)
-// }
-
-function isImageUrl(url: string) {
-  if (!url) return false;
-
-  const mdast = parser.parse(url)
-  console.log(mdast);
-  console.log(mdastToSlate(mdast as any)) // expects Root, parser returns "Node" (its actually a root in my case)
-}
-
 /**
  * Convert text to mdast -> SlateJSON, then insert into the document
  */
 function convertAndInsert(editor: ReactEditor, text: string) {
   const mdast = parser.parse(text);
   const slateNodes = mdastToSlate(mdast as any)
-  console.log(mdast)
-  // const nodes: SlateNode[] = (slateNodes[0] as any).children;
-  console.log(slateNodes)
-
-  // nodes.forEach(node => {
-  //   Transforms.insertNodes(editor, node)
-  // })
-
   Transforms.insertNodes(editor, slateNodes);
 }
-
-// const isImageUrl = url => {
-//   if (!url) return false
-//   if (!isUrl(url)) return false
-//   const ext = new URL(url).pathname.split('.').pop()
-//   return imageExtensions.includes(ext)
-// }
-// https://cdn.shopify.com/s/files/1/3106/5828/products/IMG_9385_1024x1024@2x.jpg?v=1577795595
-// const imageExtensionRegex =
-
-// Copied from this repo: https://github.com/arthurvr/image-extensions
-// Which is an npm package that is just a json file 
-const imageExtensions = [
-	"ase",
-	"art",
-	"bmp",
-	"blp",
-	"cd5",
-	"cit",
-	"cpt",
-	"cr2",
-	"cut",
-	"dds",
-	"dib",
-	"djvu",
-	"egt",
-	"exif",
-	"gif",
-	"gpl",
-	"grf",
-	"icns",
-	"ico",
-	"iff",
-	"jng",
-	"jpeg",
-	"jpg",
-	"jfif",
-	"jp2",
-	"jps",
-	"lbm",
-	"max",
-	"miff",
-	"mng",
-	"msp",
-	"nitf",
-	"ota",
-	"pbm",
-	"pc1",
-	"pc2",
-	"pc3",
-	"pcf",
-	"pcx",
-	"pdn",
-	"pgm",
-	"PI1",
-	"PI2",
-	"PI3",
-	"pict",
-	"pct",
-	"pnm",
-	"pns",
-	"ppm",
-	"psb",
-	"psd",
-	"pdd",
-	"psp",
-	"px",
-	"pxm",
-	"pxr",
-	"qfx",
-	"raw",
-	"rle",
-	"sct",
-	"sgi",
-	"rgb",
-	"int",
-	"bw",
-	"tga",
-	"tiff",
-	"tif",
-	"vtf",
-	"xbm",
-	"xcf",
-	"xpm",
-	"3dv",
-	"amf",
-	"ai",
-	"awg",
-	"cgm",
-	"cdr",
-	"cmx",
-	"dxf",
-	"e2d",
-	"egt",
-	"eps",
-	"fs",
-	"gbr",
-	"odg",
-	"svg",
-	"stl",
-	"vrml",
-	"x3d",
-	"sxd",
-	"v2d",
-	"vnd",
-	"wmf",
-	"emf",
-	"art",
-	"xar",
-	"png",
-	"webp",
-	"jxr",
-	"hdp",
-	"wdp",
-	"cur",
-	"ecw",
-	"iff",
-	"lbm",
-	"liff",
-	"nrrd",
-	"pam",
-	"pcx",
-	"pgf",
-	"sgi",
-	"rgb",
-	"rgba",
-	"bw",
-	"int",
-	"inta",
-	"sid",
-	"ras",
-	"sun",
-	"tga"
-]
